--- a/hail/python/test/hail/matrixtable/test_matrix_table.py
+++ b/hail/python/test/hail/matrixtable/test_matrix_table.py
@@ -206,6 +206,16 @@ class Tests(unittest.TestCase):
         with self.assertRaises(ValueError):
             mt.explode_rows('b')
 
+    def test_group_by_field_lifetimes(self):
+        mt = hl.utils.range_matrix_table(3, 3)
+        mt2 = (mt.group_rows_by(row_idx='100')
+               .aggregate(x=hl.agg.collect_as_set(mt.row_idx + 5)))
+        assert mt2.aggregate_entries(hl.agg.all(mt2.x == hl.set({5, 6, 7})))
+
+        mt3 = (mt.group_cols_by(col_idx='100')
+               .aggregate(x=hl.agg.collect_as_set(mt.col_idx + 5)))
+        assert mt3.aggregate_entries(hl.agg.all(mt3.x == hl.set({5, 6, 7})))
+
     def test_aggregate_cols_by(self):
         mt = hl.utils.range_matrix_table(2, 4)
         mt = (mt.annotate_cols(group=mt.col_idx < 2)