Stacked on #5103 

Fixes #5100 

This PR changes the Python select and key_by operators to generate
the IR we'd expect them to be generating (e.g. `ht.select('x')` emits
a `SelectFields` instead of a `MakeStruct`).

In the process, I found and fixed a bug in group expressions for 
`GroupedMatrixTable`. This is tested for both tables and matrix tables
in the new tests in `test_table` and `test_matrix_table`.

Some timings:

    >>> mt = hl.read_matrix_table('/Users/tpoterba/data/profile.mt')
    >>> %timeit mt.select_entries('GT')._force_count_rows()
    
master: 

    1.64 s ± 106 ms per loop
     
PR: 

    967 ms ± 61.1 ms per loop