@tpoterba 2019-01-10
This is probably even more pronounced for BGEN than MT!