@tpoterba 2019-01-10
This is a meaty PR, I'm sorry for that. But I think it'll be pretty hard to break up :(

I can't, for instance, easily fix the bugs in GroupedMatrixTable separately from the infrastructure code.