@danking 2019-01-10
Yeah. It seems there’s some clean up (check vs get, same patttern for all the annotate/select methids), then some group stuff that I haven’t unpacked, and I haven’t found what fixed the bug yet.