@tpoterba 2019-01-10
The bug was fixed by the logic in `result` that uses UIDs for computed keys, renaming them after aggregation. I'm happy to walk you through this if you want.