@danking 2019-01-17
looks like some fields (`cols`) are changing names